# src/web/openapi/running-performance.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RunningPerformance.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "operationId": "GetProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AthleteProfileResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "operationId": "UpdateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAthleteProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AthleteProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AthleteProfileResponse": {
        "required": [
          "displayName",
          "birthDate",
          "heightCm",
          "weightKg",
          "sex",
          "timezoneName",
          "locale",
          "unitSystem",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "birthDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "heightCm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "weightKg": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "sex": {
            "type": "string"
          },
          "timezoneName": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "unitSystem": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "UpdateAthleteProfileRequest": {
        "required": [
          "displayName",
          "birthDate",
          "heightCm",
          "weightKg",
          "sex",
          "timezoneName",
          "locale",
          "unitSystem"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "birthDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "heightCm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "weightKg": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "sex": {
            "type": "string"
          },
          "timezoneName": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "unitSystem": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}